feat: default dockerfile cmd to `serve`
This changes the default behavior from `help` (no command) to `serve` when
starting a container. Many servers such as postgresql, redis and more
uses a similar pattern ("sane defaults").

It is added in a separate Dockerfile to prepare testing.

Note: this will not break existing behavior since passing a command when
starting the container will override `serve`.